refactor: split classes and fix some bugs
Runner classes and components classes function differently.
Fix the bug when entering only one word.
