--- src/main/data/task_list.txt
@@ -1,2 +1,0 @@
-T0understand
-D1erre  /by 2023-09-10
