+++ src/main/java/components/Deadline.java
@@ -1,8 +1,7 @@
-package duke;
+package components;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 public class Deadline extends Task {
     protected LocalDate by;
