+++ src/main/java/components/Event.java
@@ -1,4 +1,4 @@
-package duke;
+package components;
 
 public class Event extends Task {
     protected String from;
