+++ src/main/java/runner/TaskList.java
@@ -1,4 +1,6 @@
-package duke;
+package runner;
+import components.Task;
+
 import java.util.ArrayList;
 import java.util.List;
